+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - add spaceFilter for Mrgsolve, Julia
 - version selection in #export {format: SBML}: support for L2V3, L2V4, L2V5
 - add support @TimeSwitcher as event in SimBiology (instead of doses)
+- fix bug with not a deep clone in Expression
 
 ## 0.5.7